feat: 패키지별 Encoding Service 메서드 구현
- 각 테이블의 ID 변환을 위한 서비스 메서드 구현
