+++ src/main/java/com/ssginc/wms/incomeApply/IncomeApplyService.java
@@ -1,12 +1,12 @@
 
-package com.ssginc.wms.supply;
+package com.ssginc.wms.incomeApply;
 
-public class SupplyService {
+public class IncomeApplyService {
     // int --> String
-    public static String encodeSupplyId(int supplyId) {
+    public static String encodeApplyId(int incomeApplyId) {
         StringBuilder sTemp = new StringBuilder();
-        sTemp.append("S");
-        String convertedId = String.valueOf(supplyId);
+        sTemp.append("IA");
+        String convertedId = String.valueOf(incomeApplyId);
 
         for (int i = 0; i < 4 - convertedId.length(); i++) {
             sTemp.append("0");
