+++ src/main/java/com/ssginc/wms/product/ProductService.java
@@ -12,4 +12,32 @@ public class ProductService {
 
         return result;
     }
+
+    // int --> String
+    public static String encodeProductId(int productId) {
+        StringBuilder sTemp = new StringBuilder();
+        sTemp.append("P");
+        String convertedId = String.valueOf(productId);
+
+        for (int i = 0; i < 4 - convertedId.length(); i++) {
+            sTemp.append("0");
+        }
+        sTemp.append(convertedId);
+
+        return sTemp.toString();
+    }
+
+    // int --> String
+    public static String encodeCategoryId(int categoryId) {
+        StringBuilder sTemp = new StringBuilder();
+        sTemp.append("C");
+        String convertedId = String.valueOf(categoryId);
+
+        for (int i = 0; i < 4 - convertedId.length(); i++) {
+            sTemp.append("0");
+        }
+        sTemp.append(convertedId);
+
+        return sTemp.toString();
+    }
 }
